A loot of code for orders

+++ app/src/main/java/com/example/uscdrinkdoor/Item.java
@@ -3,11 +3,11 @@ package com.example.uscdrinkdoor;
 public class Item {
     String itemName;
     String description;
-    long price;
-    long caffeineMg;
+    float price;
+    float caffeineMg;
     String sellerEmail;
 
-    public Item(String itemName, String description, long price, long caffeineMg, String email) {
+    public Item(String itemName, String description, float price, float caffeineMg, String email) {
         this.itemName = itemName;
         this.description = description;
         this.price = price;
@@ -20,9 +20,9 @@ public class Item {
 
     public String getDescription() { return description; }
 
-    public long getPrice() { return price; }
+    public float getPrice() { return price; }
 
-    public long getCaffeine() { return caffeineMg; }
+    public float getCaffeine() { return caffeineMg; }
 
     public String getSellerEmail() { return this.sellerEmail; }
